webscraper_news_docker/requirements.txt
flask==1.1.1
newspaper3k==0.2.8
requests==2.22.0
pymongo==3.9.0
apscheduler==3.6.3
